requirements.txt
requests
mysql-connector==2.1.4
flask
